[Core] Some improvements.

+++ Src/Data/Position.cs
@@ -2,7 +2,7 @@ namespace ByteCrusher.Data
 {
   public class Position
   {
-    public int X;
-    public int Y;
+    public float X;
+    public float Y;
   }
 }
